+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### **Chat Enhancements**
 **Improvements**
 - Code refactors
+- Compact chat messages have been improved to not leave behind empty separator lines without text in between them.
 
 **Fixes**
-- Fixed a bug that removed blank lines from chat
+- Fixed a bug that removed blank lines from some chat messages